Create a custom provider and label for ASR to limit number of jobs
To ensure that only a single ASR job can run at any time, this patch
adds a new nodepool provider and label with a limited max-server number.
This number can be changed to set a larger number of jobs when the jobs
are modified to allow for parallel running, and hopefully the need for
this custom provider and label will go away if parallel testing on the
ASR works in a stable way.