Add ability for certain jobs to reuse jenkins slaves
By default zuul will set a jenkins slave to offline after running the
job, this is so that nodepool nodes don't get reused before nodepool
cleans them up. However for the cinder CI and maybe other CIs in the
future we want to be able to reuse the jenkins slave because its a
static server. This change makes it so that setting a "reusenode" tag on
the job in the zuul config will tell zuul for a specific job we want to
reuse the node and so don't offline it.