Skip running the CI if a patch only contains docs or UT changes
This patch ensures that we don't run the CI jobs on patches that would
just be a waste of resources. Docs and UT patch changes aren't changing
any functional code, so there is little need to run the full CI for
those changes as UT runs and docs builds should be catching those
errors.