Add ironic-cimc-nexus jobs in experimental for networking-cisco
This patch adds a new ironic-cimc-nexus job for testing ironic with the
CIMC driver and the nexus neutron ml2 driver. In order to speed up
development this patch also adds an abilty to set a regex in the commit
message of the patch testing the new jobs so that we can skip unrelated
jobs.