Provide more control over the jenkins job names
This patch changes the jenkins job template to remove the dsvm-tempest
part and include that in the job-name field. This provides more control
over the job naming while still using the same single template for all
our jobs. As dsvm-tempest is now part of the job name, the playbooks for
each job have to have their names updated to match the change in
jenkins.